+++ ignite/cmd/scaffold.go
@@ -170,7 +170,7 @@ func scaffoldType(
 	return nil
 }
 
-func gitChangesConfirmPreRunHandler(cmd *cobra.Command, args []string) error {
+func gitChangesConfirmPreRunHandler(cmd *cobra.Command, _ []string) error {
 	// Don't confirm when the "--yes" flag is present
 	if getYes(cmd) {
 		return nil