+++ ignite/cmd/scaffold_mwasm.go
@@ -22,7 +22,7 @@ func NewScaffoldWasm() *cobra.Command {
 	return c
 }
 
-func scaffoldWasmHandler(cmd *cobra.Command, args []string) error {
+func scaffoldWasmHandler(cmd *cobra.Command, _ []string) error {
 	appPath := flagGetPath(cmd)
 
 	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))