+++ ignite/cmd/scaffold_vue.go
@@ -24,7 +24,7 @@ func NewScaffoldVue() *cobra.Command {
 	return c
 }
 
-func scaffoldVueHandler(cmd *cobra.Command, args []string) error {
+func scaffoldVueHandler(cmd *cobra.Command, _ []string) error {
 	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))
 	defer session.End()