+++ ignite/pkg/cosmosanalysis/cosmosanalysis.go
@@ -18,7 +18,7 @@ import (
 
 const (
 	cosmosModulePath     = "github.com/cosmos/cosmos-sdk"
-	tendermintModulePath = "github.com/tendermint/tendermint"
+	tendermintModulePath = "github.com/cometbft/cometbft"
 	appFileName          = "app.go"
 	defaultAppFilePath   = "app/" + appFileName
 )
@@ -27,6 +27,7 @@ var appImplementation = []string{
 	"Name",
 	"BeginBlocker",
 	"EndBlocker",
+	"RegisterAPIRoutes",
 }
 
 // implementation tracks the implementation of an interface for a given struct.