+++ ignite/pkg/cosmosclient/gasometer.go
@@ -4,7 +4,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	txtypes "github.com/cosmos/cosmos-sdk/types/tx"
-	gogogrpc "github.com/gogo/protobuf/grpc"
+	gogogrpc "github.com/cosmos/gogoproto/grpc"
 )
 
 // gasometer implements the Gasometer interface.