+++ ignite/templates/typed/simapp.go
@@ -2,6 +2,7 @@ package typed
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/ignite/cli/ignite/pkg/multiformatname"
 	"github.com/ignite/cli/ignite/pkg/placeholder"
@@ -42,6 +43,20 @@ func ModuleSimulationMsgModify(
 	%[1]v`
 		replacementOp := fmt.Sprintf(templateOp, PlaceholderSimappOperation, msg, typeName.UpperCamel, moduleName)
 		content = replacer.Replace(content, PlaceholderSimappOperation, replacementOp)
+
+		if strings.Contains(content, PlaceholderSimappOperationMsg) {
+			templateOpMsg := `simulation.NewWeightedProposalMsg(
+	opWeightMsg%[2]v%[3]v,
+	defaultWeightMsg%[2]v%[3]v,
+	func(r *rand.Rand, ctx sdk.Context, accs []simtypes.Account) sdk.Msg {
+		%[4]vsimulation.SimulateMsg%[2]v%[3]v(am.accountKeeper, am.bankKeeper, am.keeper)
+		return nil
+	},
+),
+%[1]v`
+			replacementOpMsg := fmt.Sprintf(templateOpMsg, PlaceholderSimappOperationMsg, msg, typeName.UpperCamel, moduleName)
+			content = replacer.Replace(content, PlaceholderSimappOperationMsg, replacementOpMsg)
+		}
 	}
 	return content
 }