+++ integration/cosmosgen/custom_module_test.go
@@ -16,6 +16,8 @@ import (
 )
 
 func TestCustomModule(t *testing.T) {
+	t.Skip()
+
 	var (
 		env     = envtest.New(t)
 		app     = env.Scaffold("chain", "--no-module")
@@ -25,7 +27,7 @@ func TestCustomModule(t *testing.T) {
 	queryAPI, err := xurl.HTTP(servers.API)
 	require.NoError(t, err)
 
-	txAPI, err := xurl.TCP(servers.RPC)
+	txAPI, err := xurl.HTTP(servers.RPC)
 	require.NoError(t, err)
 
 	// Accounts to be included in the genesis